+++ client/src/styles/Sidebar.styles.js
@@ -1,7 +1,10 @@
 import styled from "styled-components"
+import { media } from "../styles/theme/mixins"
 
 export const SidebarStyles = styled.div`
   width: ${props => `${props.theme.sidebarWidth}px`};
+  ${media.tablet`display: none;`}
+
   button {
     width: 100%;
     border-radius: 0;
